feat: organized implementation code in a new directory
